praxiscore-hub-net/src/main/java/org/praxislive/hub/net/IonEncoder.java
/*
 * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
 *
 * Copyright 2023 Neil C Smith.
 *
 * This code is free software; you can redistribute it and/or modify it
 * under the terms of the GNU Lesser General Public License version 3 only, as
 * published by the Free Software Foundation.
 *
 * This code is distributed in the hope that it will be useful, but WITHOUT
 * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
 * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Lesser General Public License
 * version 3 for more details.
 *
 * You should have received a copy of the GNU Lesser General Public License version 3
 * along with this work; if not, see http://www.gnu.org/licenses/
 *
 *
 * Please visit https://www.praxislive.org if you need additional information or
 * have any questions.
 */
package org.praxislive.hub.net;

import io.netty.buffer.ByteBuf;
import io.netty.buffer.ByteBufOutputStream;
import io.netty.channel.ChannelHandlerContext;
import io.netty.handler.codec.MessageToByteEncoder;
import java.util.List;

class IonEncoder extends MessageToByteEncoder<List<Message>> {

    @Override
    protected void encode(ChannelHandlerContext ctx, List<Message> messages, ByteBuf out) throws Exception {
        int start = out.writerIndex();

        try (var stream = new ByteBufOutputStream(out)) {
            stream.writeInt(0);
            IonCodec.getDefault().writeMessages(messages, stream);
        }

        int end = out.writerIndex();

        out.setInt(start, end - start - 4);

    }

}
